Move shared coded to Shared, update
