+++ src/Servers/Kestrel/Core/src/Internal/Http/HttpHeaders.Generated.cs
@@ -5768,7 +5768,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
         }
         
         [MethodImpl(MethodImplOptions.AggressiveOptimization)]
-        public unsafe void Append(Span<byte> name, Span<byte> value)
+        public unsafe void Append(ReadOnlySpan<byte> name, ReadOnlySpan<byte> value)
         {
             ref byte nameStart = ref MemoryMarshal.GetReference(name);
             ref StringValues values = ref Unsafe.AsRef<StringValues>(null);
