--- src/Servers/Kestrel/Core/src/Internal/Http2/HPack/HPackDecoder.cs
@@ -1,472 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.Buffers;
-using System.Diagnostics;
-using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http;
-
-namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2.HPack
-{
-    internal class HPackDecoder
-    {
-        private enum State
-        {
-            Ready,
-            HeaderFieldIndex,
-            HeaderNameIndex,
-            HeaderNameLength,
-            HeaderNameLengthContinue,
-            HeaderName,
-            HeaderValueLength,
-            HeaderValueLengthContinue,
-            HeaderValue,
-            DynamicTableSizeUpdate
-        }
-
-        public const int DefaultHeaderTableSize = 4096;
-        public const int DefaultStringOctetsSize = 4096;
-        public const int DefaultMaxResponseHeadersLength = 64 * 1024;
-
-        // http://httpwg.org/specs/rfc7541.html#rfc.section.6.1
-        //   0   1   2   3   4   5   6   7
-        // +---+---+---+---+---+---+---+---+
-        // | 1 |        Index (7+)         |
-        // +---+---------------------------+
-        private const byte IndexedHeaderFieldMask = 0x80;
-        private const byte IndexedHeaderFieldRepresentation = 0x80;
-
-        // http://httpwg.org/specs/rfc7541.html#rfc.section.6.2.1
-        //   0   1   2   3   4   5   6   7
-        // +---+---+---+---+---+---+---+---+
-        // | 0 | 1 |      Index (6+)       |
-        // +---+---+-----------------------+
-        private const byte LiteralHeaderFieldWithIncrementalIndexingMask = 0xc0;
-        private const byte LiteralHeaderFieldWithIncrementalIndexingRepresentation = 0x40;
-
-        // http://httpwg.org/specs/rfc7541.html#rfc.section.6.2.2
-        //   0   1   2   3   4   5   6   7
-        // +---+---+---+---+---+---+---+---+
-        // | 0 | 0 | 0 | 0 |  Index (4+)   |
-        // +---+---+-----------------------+
-        private const byte LiteralHeaderFieldWithoutIndexingMask = 0xf0;
-        private const byte LiteralHeaderFieldWithoutIndexingRepresentation = 0x00;
-
-        // http://httpwg.org/specs/rfc7541.html#rfc.section.6.2.3
-        //   0   1   2   3   4   5   6   7
-        // +---+---+---+---+---+---+---+---+
-        // | 0 | 0 | 0 | 1 |  Index (4+)   |
-        // +---+---+-----------------------+
-        private const byte LiteralHeaderFieldNeverIndexedMask = 0xf0;
-        private const byte LiteralHeaderFieldNeverIndexedRepresentation = 0x10;
-
-        // http://httpwg.org/specs/rfc7541.html#rfc.section.6.3
-        //   0   1   2   3   4   5   6   7
-        // +---+---+---+---+---+---+---+---+
-        // | 0 | 0 | 1 |   Max size (5+)   |
-        // +---+---------------------------+
-        private const byte DynamicTableSizeUpdateMask = 0xe0;
-        private const byte DynamicTableSizeUpdateRepresentation = 0x20;
-
-        // http://httpwg.org/specs/rfc7541.html#rfc.section.5.2
-        //   0   1   2   3   4   5   6   7
-        // +---+---+---+---+---+---+---+---+
-        // | H |    String Length (7+)     |
-        // +---+---------------------------+
-        private const byte HuffmanMask = 0x80;
-
-        private const int IndexedHeaderFieldPrefix = 7;
-        private const int LiteralHeaderFieldWithIncrementalIndexingPrefix = 6;
-        private const int LiteralHeaderFieldWithoutIndexingPrefix = 4;
-        private const int LiteralHeaderFieldNeverIndexedPrefix = 4;
-        private const int DynamicTableSizeUpdatePrefix = 5;
-        private const int StringLengthPrefix = 7;
-
-        private readonly int _maxDynamicTableSize;
-        private readonly int _maxResponseHeadersLength;
-        private readonly DynamicTable _dynamicTable;
-        private readonly IntegerDecoder _integerDecoder = new IntegerDecoder();
-        private byte[] _stringOctets;
-        private byte[] _headerNameOctets;
-        private byte[] _headerValueOctets;
-
-        private State _state = State.Ready;
-        private byte[] _headerName;
-        private int _stringIndex;
-        private int _stringLength;
-        private int _headerNameLength;
-        private int _headerValueLength;
-        private bool _index;
-        private bool _huffman;
-        private bool _headersObserved;
-
-        public HPackDecoder(int maxDynamicTableSize = DefaultHeaderTableSize, int maxResponseHeadersLength = DefaultMaxResponseHeadersLength)
-            : this(maxDynamicTableSize, maxResponseHeadersLength, new DynamicTable(maxDynamicTableSize)) 
-        {
-        }
-
-        // For testing.
-        internal HPackDecoder(int maxDynamicTableSize, int maxResponseHeadersLength, DynamicTable dynamicTable)
-        {
-            _maxDynamicTableSize = maxDynamicTableSize;
-            _maxResponseHeadersLength = maxResponseHeadersLength;
-            _dynamicTable = dynamicTable;
-
-            _stringOctets = new byte[maxResponseHeadersLength];
-            _headerNameOctets = new byte[maxResponseHeadersLength];
-            _headerValueOctets = new byte[maxResponseHeadersLength];
-        }
-
-        public void Decode(in ReadOnlySequence<byte> data, bool endHeaders, IHttpHeadersHandler handler)
-        {
-            foreach (var segment in data)
-            {
-                var span = segment.Span;
-                for (var i = 0; i < span.Length; i++)
-                {
-                    OnByte(span[i], handler);
-                }
-            }
-
-            if (endHeaders)
-            {
-                if (_state != State.Ready)
-                {
-                    throw new HPackDecodingException(CoreStrings.HPackErrorIncompleteHeaderBlock);
-                }
-
-                _headersObserved = false;
-            }
-        }
-
-        private void OnByte(byte b, IHttpHeadersHandler handler)
-        {
-            int intResult;
-            switch (_state)
-            {
-                case State.Ready:
-                    // TODO: Instead of masking and comparing each prefix value,
-                    // consider doing a 16-way switch on the first four bits (which is the max prefix size).
-                    // Look at this once we have more concrete perf data.
-                    if ((b & IndexedHeaderFieldMask) == IndexedHeaderFieldRepresentation)
-                    {
-                        _headersObserved = true;
-
-                        int val = b & ~IndexedHeaderFieldMask;
-
-                        if (_integerDecoder.BeginTryDecode((byte)val, IndexedHeaderFieldPrefix, out intResult))
-                        {
-                            OnIndexedHeaderField(intResult, handler);
-                        }
-                        else
-                        {
-                            _state = State.HeaderFieldIndex;
-                        }
-                    }
-                    else if ((b & LiteralHeaderFieldWithIncrementalIndexingMask) == LiteralHeaderFieldWithIncrementalIndexingRepresentation)
-                    {
-                        _headersObserved = true;
-
-                        _index = true;
-                        int val = b & ~LiteralHeaderFieldWithIncrementalIndexingMask;
-
-                        if (val == 0)
-                        {
-                            _state = State.HeaderNameLength;
-                        }
-                        else if (_integerDecoder.BeginTryDecode((byte)val, LiteralHeaderFieldWithIncrementalIndexingPrefix, out intResult))
-                        {
-                            OnIndexedHeaderName(intResult);
-                        }
-                        else
-                        {
-                            _state = State.HeaderNameIndex;
-                        }
-                    }
-                    else if ((b & LiteralHeaderFieldWithoutIndexingMask) == LiteralHeaderFieldWithoutIndexingRepresentation)
-                    {
-                        _headersObserved = true;
-
-                        _index = false;
-                        int val = b & ~LiteralHeaderFieldWithoutIndexingMask;
-
-                        if (val == 0)
-                        {
-                            _state = State.HeaderNameLength;
-                        }
-                        else if (_integerDecoder.BeginTryDecode((byte)val, LiteralHeaderFieldWithoutIndexingPrefix, out intResult))
-                        {
-                            OnIndexedHeaderName(intResult);
-                        }
-                        else
-                        {
-                            _state = State.HeaderNameIndex;
-                        }
-                    }
-                    else if ((b & LiteralHeaderFieldNeverIndexedMask) == LiteralHeaderFieldNeverIndexedRepresentation)
-                    {
-                        _headersObserved = true;
-
-                        _index = false;
-                        int val = b & ~LiteralHeaderFieldNeverIndexedMask;
-
-                        if (val == 0)
-                        {
-                            _state = State.HeaderNameLength;
-                        }
-                        else if (_integerDecoder.BeginTryDecode((byte)val, LiteralHeaderFieldNeverIndexedPrefix, out intResult))
-                        {
-                            OnIndexedHeaderName(intResult);
-                        }
-                        else
-                        {
-                            _state = State.HeaderNameIndex;
-                        }
-                    }
-                    else if ((b & DynamicTableSizeUpdateMask) == DynamicTableSizeUpdateRepresentation)
-                    {
-                        // https://tools.ietf.org/html/rfc7541#section-4.2
-                        // This dynamic table size
-                        // update MUST occur at the beginning of the first header block
-                        // following the change to the dynamic table size.
-                        if (_headersObserved)
-                        {
-                            throw new HPackDecodingException(CoreStrings.HPackErrorDynamicTableSizeUpdateNotAtBeginningOfHeaderBlock);
-                        }
-
-                        if (_integerDecoder.BeginTryDecode((byte)(b & ~DynamicTableSizeUpdateMask), DynamicTableSizeUpdatePrefix, out intResult))
-                        {
-                            SetDynamicHeaderTableSize(intResult);
-                        }
-                        else
-                        {
-                            _state = State.DynamicTableSizeUpdate;
-                        }
-                    }
-                    else
-                    {
-                        // Can't happen
-                        Debug.Fail("Unreachable code");
-                        throw new InvalidOperationException();
-                    }
-
-                    break;
-                case State.HeaderFieldIndex:
-                    if (_integerDecoder.TryDecode(b, out intResult))
-                    {
-                        OnIndexedHeaderField(intResult, handler);
-                    }
-
-                    break;
-                case State.HeaderNameIndex:
-                    if (_integerDecoder.TryDecode(b, out intResult))
-                    {
-                        OnIndexedHeaderName(intResult);
-                    }
-
-                    break;
-                case State.HeaderNameLength:
-                    _huffman = (b & HuffmanMask) != 0;
-
-                    if (_integerDecoder.BeginTryDecode((byte)(b & ~HuffmanMask), StringLengthPrefix, out intResult))
-                    {
-                        if (intResult == 0)
-                        {
-                            // TODO synchronize Http2 resource strings.
-                            throw new HPackDecodingException(/*SR.Format(SR.net_http_invalid_response_header_name, "")*/);
-                        }
-
-                        OnStringLength(intResult, nextState: State.HeaderName);
-                    }
-                    else
-                    {
-                        _state = State.HeaderNameLengthContinue;
-                    }
-
-                    break;
-                case State.HeaderNameLengthContinue:
-                    if (_integerDecoder.TryDecode(b, out intResult))
-                    {
-                        // IntegerDecoder disallows overlong encodings, where an integer is encoded with more bytes than is strictly required.
-                        // 0 should always be represented by a single byte, so we shouldn't need to check for it in the continuation case.
-                        Debug.Assert(intResult != 0, "A header name length of 0 should never be encoded with a continuation byte.");
-
-                        OnStringLength(intResult, nextState: State.HeaderName);
-                    }
-
-                    break;
-                case State.HeaderName:
-                    _stringOctets[_stringIndex++] = b;
-
-                    if (_stringIndex == _stringLength)
-                    {
-                        OnString(nextState: State.HeaderValueLength);
-                    }
-
-                    break;
-                case State.HeaderValueLength:
-                    _huffman = (b & HuffmanMask) != 0;
-
-                    if (_integerDecoder.BeginTryDecode((byte)(b & ~HuffmanMask), StringLengthPrefix, out intResult))
-                    {
-                        OnStringLength(intResult, nextState: State.HeaderValue);
-
-                        if (intResult == 0)
-                        {
-                            ProcessHeaderValue(handler);
-                        }
-                    }
-                    else
-                    {
-                        _state = State.HeaderValueLengthContinue;
-                    }
-
-                    break;
-                case State.HeaderValueLengthContinue:
-                    if (_integerDecoder.TryDecode(b, out intResult))
-                    {
-                        // IntegerDecoder disallows overlong encodings where an integer is encoded with more bytes than is strictly required.
-                        // 0 should always be represented by a single byte, so we shouldn't need to check for it in the continuation case.
-                        Debug.Assert(intResult != 0, "A header value length of 0 should never be encoded with a continuation byte.");
-
-                        OnStringLength(intResult, nextState: State.HeaderValue);
-                    }
-
-                    break;
-                case State.HeaderValue:
-                    _stringOctets[_stringIndex++] = b;
-
-                    if (_stringIndex == _stringLength)
-                    {
-                        ProcessHeaderValue(handler);
-                    }
-
-                    break;
-                case State.DynamicTableSizeUpdate:
-                    if (_integerDecoder.TryDecode(b, out intResult))
-                    {
-                        SetDynamicHeaderTableSize(intResult);
-                        _state = State.Ready;
-                    }
-
-                    break;
-                default:
-                    // Can't happen
-                    Debug.Fail("HPACK decoder reach an invalid state");
-                    throw new InvalidOperationException();
-            }
-        }
-
-        private void ProcessHeaderValue(IHttpHeadersHandler handler)
-        {
-            OnString(nextState: State.Ready);
-
-            var headerNameSpan = new Span<byte>(_headerName, 0, _headerNameLength);
-            var headerValueSpan = new Span<byte>(_headerValueOctets, 0, _headerValueLength);
-
-            handler.OnHeader(headerNameSpan, headerValueSpan);
-
-            if (_index)
-            {
-                _dynamicTable.Insert(headerNameSpan, headerValueSpan);
-            }
-        }
-
-        private void OnIndexedHeaderField(int index, IHttpHeadersHandler handler)
-        {
-            HeaderField header = GetHeader(index);
-            handler.OnHeader(new Span<byte>(header.Name), new Span<byte>(header.Value));
-            _state = State.Ready;
-        }
-
-        private void OnIndexedHeaderName(int index)
-        {
-            HeaderField header = GetHeader(index);
-            _headerName = header.Name;
-            _headerNameLength = header.Name.Length;
-            _state = State.HeaderValueLength;
-        }
-
-        private void OnStringLength(int length, State nextState)
-        {
-            if (length > _stringOctets.Length)
-            {
-                if (length > _maxResponseHeadersLength)
-                {
-                    throw new HPackDecodingException(CoreStrings.FormatHPackStringLengthTooLarge(length, _stringOctets.Length));
-                }
-
-                _stringOctets = new byte[Math.Max(length, _stringOctets.Length * 2)];
-            }
-
-            _stringLength = length;
-            _stringIndex = 0;
-            _state = nextState;
-        }
-
-        private void OnString(State nextState)
-        {
-            int Decode(ref byte[] dst)
-            {
-                if (_huffman)
-                {
-                    return Huffman.Decode(new ReadOnlySpan<byte>(_stringOctets, 0, _stringLength), ref dst);
-                }
-                else
-                {
-                    if (dst.Length < _stringLength)
-                    {
-                        dst = new byte[Math.Max(_stringLength, dst.Length * 2)];
-                    }
-
-                    Buffer.BlockCopy(_stringOctets, 0, dst, 0, _stringLength);
-                    return _stringLength;
-                }
-            }
-
-            try
-            {
-                if (_state == State.HeaderName)
-                {
-                    _headerNameLength = Decode(ref _headerNameOctets);
-                    _headerName = _headerNameOctets;
-                }
-                else
-                {
-                    _headerValueLength = Decode(ref _headerValueOctets);
-                }
-            }
-            catch (HuffmanDecodingException ex)
-            {
-                throw new HPackDecodingException(CoreStrings.HPackHuffmanError, ex);
-            }
-
-            _state = nextState;
-        }
-
-        private HeaderField GetHeader(int index)
-        {
-            try
-            {
-                return index <= StaticTable.Count
-                    ? StaticTable.Get(index - 1)
-                    : _dynamicTable[index - StaticTable.Count - 1];
-            }
-            catch (IndexOutOfRangeException ex)
-            {
-                // Header index out of range.
-                throw new HPackDecodingException(CoreStrings.FormatHPackErrorIndexOutOfRange(index), ex);
-            }
-        }
-
-        private void SetDynamicHeaderTableSize(int size)
-        {
-            if (size > _maxDynamicTableSize)
-            {
-                throw new HPackDecodingException(CoreStrings.FormatHPackErrorDynamicTableSizeUpdateTooLarge(size, _maxDynamicTableSize));
-            }
-
-            _dynamicTable.Resize(size);
-        }
-    }
-}
