+++ src/Servers/Kestrel/Core/src/SR.cs
@@ -1,0 +1,13 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace System.Net.Http
+{
+    internal static partial class SR
+    {
+        internal static string Format(string resourceFormat, params object[] args)
+        {
+            return resourceFormat;
+        }
+    }
+}
