+++ src/Shared/Http2/IHttpHeadersHandler.cs
@@ -1,0 +1,12 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Net.Http
+{
+    internal interface IHttpHeadersHandler
+    {
+        void OnHeader(ReadOnlySpan<byte> name, ReadOnlySpan<byte> value);
+        void OnHeadersComplete(bool endStream);
+    }
+}
